rename folders and render blank html
