+++ app/pages/home/index.html
@@ -1,0 +1,9 @@
+<html>
+
+<head>
+    <title>chatroom</title>
+</head>
+
+<body>OKAY</body>
+
+</html>
